removed all pinvoke methods.
one new class added called OpCode to provide the RDTSC CPU instructions which give the timestamp of the CPU, this is very important for ZEROMQ performance.
the OpCode use native methods of windows and Mono.Posix on unix and linux platforms, on other platform this feature is not supported but the library should compile and work.