Proposal for suspended poller operations
Ran the tests included in the solution, and it seemed to perform as
expected. Ran a few tests of my own, and it did alright. The performance
impact for scheduled AddSockets and RemoveSockets looks minimal! For
unscheduled use, it performs well unless the operations are
massively-parallelized, in which case it performs as well as
ConcurrentQueue does.