Use a stable hash function for computing IPC port
Fixes #1080

The previous code used String.GetHashCode, however in .NET 5 and later
this value is not stable across process runs. The value intentionally
varies due to "hash randomization", a security feature.

In this case, we need a stable value, and we need the same value
irrespective of the version of .NET or the platform.

To ensure compatibility, we take the algorithm from .NET Framework.